webapp/requirements.txt
Flask==2.2.2
Flask-Login==0.6.2
psycopg2-binary==2.9.6
Werkzeug==2.2.2
python-dotenv
pandas